+++ src/app/clubs/[id]/page.jsx
@@ -9,7 +9,7 @@ import ClubSocials from "components/clubs/ClubSocials";
 import EventsGrid from "components/events/EventsGrid";
 import Icon from "components/Icon";
 import ButtonLink from "components/Link";
-import Markdown from "components/Markdown";
+import Markdown from "components/markdown/MarkdownRenderer";
 import MembersGrid from "components/members/MembersGrid";
 import { getClub } from "utils/fetchData";
 import { stripMarkdown } from "utils/markdown";
@@ -131,9 +131,8 @@ export default async function Club(props) {
               variant="none"
               color="secondary"
               component={ButtonLink}
-              href={`/${
-                club?.category == "body" ? "student-bodies" : "clubs"
-              }/${id}/members`}
+              href={`/${club?.category == "body" ? "student-bodies" : "clubs"
+                }/${id}/members`}
             >
               <Typography
                 variant="button"
